detection_ai/requirements.txt
opencv-python-headless
numpy
ultralytics
Pillow
